pyproject.toml
[tool.poetry]
name = "repl-nix-light-photons-of-sound"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "repl_nix_light_photons_of_sound"}]

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.3.2"
openai = "^1.23.1"
google-generativeai = "^0.5.2"
requests = "^2.31.0"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
